+++ src/pages/CreateGroupPage/CreateGroupPage.styled.js
